Unify pool configurations flow and simplify code
