+++ server/package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@sv2-ui/shared": "*",
+    "bs58check": "^4.0.0",
     "cors": "^2.8.5",
     "dockerode": "^4.0.2",
     "express": "^4.18.2",
